Extract method to iterate possible pairs/triplets/etc into separate class
